Move ICalParser to util/VParser
